+++ spimdisasm/mips/symbols/MipsSymbolFunction.py
@@ -427,6 +427,8 @@ class SymbolFunction(SymbolText):
             labelSym.referenceCounter += 1
             labelSym.referenceFunctions.add(self.contextSym)
             labelSym.parentFunction = self.contextSym
+            if self.parent is not None:
+                labelSym.parentFileName = self.parent.getName()
             self.contextSym.branchLabels.add(labelSym.vram, labelSym)
 
         # Function calls