+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Avoid emitting "global" visibility on labels.
+
 ## [1.28.0] - 2024-08-09
 
 ### Added